+++ test/lib/validation/schema/__helper__/project.js
@@ -23,7 +23,7 @@ export default {
 		customConfiguration.defineTests(test, assertValidation, type);
 
 		// version specific tests
-		["2.6", "2.5", "2.4", "2.3", "2.2", "2.1", "2.0"].forEach((specVersion) => {
+		["3.0", "2.6", "2.5", "2.4", "2.3", "2.2", "2.1", "2.0"].forEach((specVersion) => {
 			// tests for all kinds and version 2.0 and above
 			test(`${type} (specVersion ${specVersion}): No metadata`, async (t) => {
 				await assertValidation(t, {
