+++ test/lib/validation/schema/specVersion/kind/extension/server-middleware.js
@@ -1,10 +1,10 @@
 import test from "ava";
 import Ajv from "ajv";
 import ajvErrors from "ajv-errors";
-import AjvCoverage from "../../../../../../../utils/AjvCoverage.js";
-import {_Validator as Validator} from "../../../../../../../../lib/validation/validator.js";
-import ValidationError from "../../../../../../../../lib/validation/ValidationError.js";
-import extension from "../../../../__helper__/extension.js";
+import AjvCoverage from "../../../../../../utils/AjvCoverage.js";
+import {_Validator as Validator} from "../../../../../../../lib/validation/validator.js";
+import ValidationError from "../../../../../../../lib/validation/ValidationError.js";
+import extension from "../../../__helper__/extension.js";
 
 async function assertValidation(t, config, expectedErrors = undefined) {
 	const validation = t.context.validator.validate({config, project: {id: "my-project"}});
@@ -25,12 +25,12 @@ async function assertValidation(t, config, expectedErrors = undefined) {
 test.before((t) => {
 	t.context.validator = new Validator({Ajv, ajvErrors});
 	t.context.ajvCoverage = new AjvCoverage(t.context.validator.ajv, {
-		includes: ["schema/specVersion/kind/extension/task.json"]
+		includes: ["schema/specVersion/kind/extension/server-middleware.json"]
 	});
 });
 
 test.after.always((t) => {
-	t.context.ajvCoverage.createReport("html", {dir: "coverage/ajv-extension-task"});
+	t.context.ajvCoverage.createReport("html", {dir: "coverage/ajv-extension-server-middleware"});
 	const thresholds = {
 		statements: 70,
 		branches: 55,
@@ -41,9 +41,9 @@ test.after.always((t) => {
 });
 
 const additionalConfiguration = {
-	"task": {
+	"middleware": {
 		"path": "/foo"
 	}
 };
 
-extension.defineTests(test, assertValidation, "task", additionalConfiguration);
+extension.defineTests(test, assertValidation, "server-middleware", additionalConfiguration);
